requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
alembic==1.14.0
email-validator==2.2.0
python-multipart==0.0.18

celery==5.3.6
redis==5.0.1
pgvector==0.2.4
google-genai==0.3.0
